+++ packages/g-plugin-device-renderer/src/drawcalls/Mesh.ts
@@ -3,7 +3,7 @@ import { Shape } from '@antv/g-lite';
 import type { Mesh } from '../Mesh';
 import { Instanced } from './Instanced';
 export class MeshDrawcall extends Instanced {
-  protected mergeXYZIntoModelMatrix = true;
+  protected mergeXYZIntoModelMatrix = false;
 
   shouldMerge(object: DisplayObject, index: number) {
     const shouldMerge = super.shouldMerge(object, index);