+++ packages/g-dom-mutation-observer-api/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @antv/g-dom-mutation-observer-api
 
+## 2.0.2
+
+### Patch Changes
+
+-   Updated dependencies [2948b0f8]
+    -   @antv/g-lite@2.0.2
+
 ## 2.0.1
 
 ### Patch Changes