+++ packages/g-gesture/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @antv/g-gesture
 
+## 3.0.2
+
+### Patch Changes
+
+-   Updated dependencies [2948b0f8]
+    -   @antv/g-lite@2.0.2
+
 ## 3.0.1
 
 ### Patch Changes