+++ packages/g-mobile-canvas-element/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @antv/g-mobile-canvas-element
 
+## 1.0.2
+
+### Patch Changes
+
+-   Updated dependencies [2948b0f8]
+    -   @antv/g-lite@2.0.2
+
 ## 1.0.1
 
 ### Patch Changes