+++ packages/g-mobile-canvas/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @antv/g-mobile-canvas
 
+## 1.0.2
+
+### Patch Changes
+
+-   Updated dependencies [2948b0f8]
+    -   @antv/g-lite@2.0.2
+    -   @antv/g-plugin-canvas-path-generator@2.0.2
+    -   @antv/g-plugin-canvas-picker@2.0.2
+    -   @antv/g-plugin-canvas-renderer@2.0.2
+    -   @antv/g-plugin-dragndrop@2.0.2
+    -   @antv/g-plugin-gesture@2.0.2
+    -   @antv/g-plugin-image-loader@2.0.2
+    -   @antv/g-plugin-mobile-interaction@1.0.2
+
 ## 1.0.1
 
 ### Patch Changes