+++ packages/g-plugin-svg-renderer/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @antv/g-plugin-svg-renderer
 
+## 2.0.3
+
+### Patch Changes
+
+-   Updated dependencies [2948b0f8]
+    -   @antv/g-lite@2.0.2
+
 ## 2.0.2
 
 ### Patch Changes